Change int to float

+++ Exercice4.java
@@ -4,17 +4,17 @@ class Exercice4 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
         System.out.println("Veuillez saisir le resultat sur 20 d'Examen1:");
-        int ex1 = sc.nextInt();
+        float ex1 = sc.nextFloat();
         System.out.println("Veuillez saisir la ponderation (%) d'Examen1:");
-        int ex1p = sc.nextInt();
+        float ex1p = sc.nextFloat();
         System.out.println("Veuillez saisir le resultat sur 20 d'Examen2:");
-        int ex2 = sc.nextInt();
+        float ex2 = sc.nextFloat();
         System.out.println("Veuillez saisir la ponderation (%) d'Examen2:");
-        int ex2p = sc.nextInt();
+        float ex2p = sc.nextFloat();
         System.out.println("Veuillez saisir le resultat sur 20 d'Examen3:");
-        int ex3 = sc.nextInt();
+        float ex3 = sc.nextFloat();
         System.out.println("Veuillez saisir la ponderation (%) d'Examen3:");
-        int ex3p = sc.nextInt();
+        float ex3p = sc.nextFloat();
         System.out.println("y = " + y);
